Refactor code to create telemetry service for Snowpipe Streaming
