+++ src/main/java/com/snowflake/kafka/connector/internal/SnowflakeConnectionServiceFactory.java
@@ -12,7 +12,7 @@ public class SnowflakeConnectionServiceFactory {
     return new SnowflakeConnectionServiceBuilder();
   }
 
-  public static class SnowflakeConnectionServiceBuilder extends EnableLogging {
+  public static class SnowflakeConnectionServiceBuilder implements SFLogger {
     private Properties prop;
     private Properties proxyProperties;
     private SnowflakeURL url;
