+++ src/test/java/com/snowflake/kafka/connector/internal/ResetProxyConfigExec.java
@@ -2,7 +2,7 @@ package com.snowflake.kafka.connector.internal;
 
 import net.snowflake.client.jdbc.SnowflakeSQLException;
 
-public class ResetProxyConfigExec extends EnableLogging {
+public class ResetProxyConfigExec implements SFLogger {
   public static void main(String[] args) throws SnowflakeSQLException {
     System.out.println("ResetProxyConfigExec::Start wiping Proxy config");
     TestUtils.resetProxyParametersInJVM();
